CMakeLists.txt changes to accomodate changing engine/ooferror2* to engine/ooferror*.
CMake requires that all targets have different names, even if
they're in different directories.  I added a TARGET_SFX option
to the swig_sources macro.  It adds the given suffix to the
cmake target.  The modified target name doesn't appear to the user
or programmer at all.  It does change the name of the shared C++
library containing the compiled swig code for the module.