Reorganized fiddlenodes to make it friendly for iterators.
The list of nodes being operated on is now retained by the
FiddleNodes object, not the FiddleNodesTargets object.  It is
passed back in to the target object's __call__ method, so the
method has the option of returning the same nodes as before, or
returning a new set, which may be given by an iterator or generator.

FiddleNodes.coreProcess() turns the iterator into a list and shuffles
it.  It passes the *shuffled list* back to the FiddleNodesTargets.
This is required in order to preserve compatibility with the previous
version and to not break the tests.