Moved the Python shuffle function in crandom.swg to C++.
This might speed up some skeleton operations, but mostly I did it
because it was easy.  A Python list object is passed in as a
PyObject*, and shuffled with Python API calls (using the same algorithm
as the C++ oofshuffle function, but different code).  The list
isn't converted to a C++ std::vector and the objects in the list
are treated as opaque PyObject*s.

Augmented the shuffle test in fundamental_tests to check that
the new shuffling method wasn't leaking reference counts.