Improvements to iterating over skeleton elements.
There is a new SkeletonElementIterator class that has the
advantage that it report how far through the list it is, so
that progress bars can be drawn properly.  By default it
iterates over all elements, but it can take an argument
that limits it to a subset.  So for example, Skeleton.activeElements
just returns SkeletonElementIterator(self, lambda e: e.active(self)).

Got rid of Skeleton.getIllegalElements(), which was identical
to Skeleton.illegalElements().

Added a missing __iter__ to ElementCornerNodeIterator.