More changes related to iterators, some more substantial than others.
The draw_elements() function in skeletoninfodisplay.py is now careful
not to create a CanvasLayer if its element iterator argument doesn't
return any elements.  This was causing the pdf output test to fail.

In debug mode the canvas can be saved in "datadump" format.  This
doesn't really relate to iterators but was useful to debug the above
problem.

Used simpler iterator code in csubproblem.spy, in a spot that
doesn't matter at all.

Did *not* remove a list conversion in element.swg, because the
result is passed to flatten1, which can't handle an iterator.