TODOs done for simplifications in Equation handling.
allEquations, defined equation.spy, is now a generator function.
It used to be an object that sort of mimicked a list, but it was
only used in contexts where a generator function would work.  Just
use "for eqn in allEquations():" instead of "for eqn in allEquations:"

There's no need to route getEquation(name) through a new swigged
getEquationByName() function.  The static C++ Equation::getEquation
can be swigged directly.