Changed FieldIndex::components to FieldIndex::getComponents.  Same for FieldIterator.
This isn't strictly necessary but will reduce confusion.  I'm going
to add a Field::components() mehtod that returns an iterator for
looping over the compoenets (FieldIndex objects) of a Field.  It'll
work for Fluxes and Equations too.

getComponents is used in only one spot in the program and is
possibly unnecessary. It's defined as a pure virtual method
in the base FieldIndex class, and is implemented in all the
derived classes, but is actually used in only one of them.