Progress on replacing FieldIterator and FieldIteratorP with real iterators.
I haven't looked at the python side yet, and the C++ side doesn't
compile, but there are enough changes that they should be committed.
It's all on a branch in case it doesn't work out.

FieldIndex is now derived from PythonExportable.

FieldIterator, which was derived from FieldIndex, is gone.

Fields will have a components(Planarity) method that returns
a Components object, which has derived classes with begin()
and end() methods that return iterators for the given Field
(or Flux or Equation) and Planarity.

Currently the Components classes and their iterators are in place,
to a first approximation.  Things don't compile yet.  Partly
because there are other things derived from FieldIndex, used
for indexing various Outputs, and they all need to be updated.