Use modern iteration techniques instead of FieldIterator and IteratorP.
All loops over field, flux, and equation components in C++ have
been replaced with range-based for loops that use the Components
mechanism implemented in fieldindex.*.  For example,
  for(IndexP i : field->components(ALL_INDICES)) {
      // do something with generic index i
  }

Some loops over symmetric tensor components are not tied to a
particular field or flux, or also require non-generic methods in
the SymTensorIndex class, so it's not sufficient to use generic
FieldIndex objects from components().  I added a symTensorIJIterator
that returns real SymTensorIndex objects.  It loops over a
SymTensorIJComponents object that is defined in fieldindex.h.
This allows
  for(SymTensorIndex ij : symTensorIJComponents) {
     ...
  }

The modified C++ files all compile.  I don't know if they work,
because the swig and python files need to be updated before everything
will build.

I deleted a bunch of 3D code in #if DIM==3 blocks to make the
files easier to read. OOF3D will never be merged into OOF2.