Updated some python loops to use new Field.components() methods.
Got rid of all iterator_all() calls in python, and most mentions
of IteratorP and IndexP, which aren't used in python anymore.

Swigged SymTensorIJComponents and gave it an __iter__ method.

Replaced the operator[](IndexP&) methods in OutputVal and OutputValue
classes with operator[](FieldIndex&) methods.  Because IndexP can
be implicitly converted to FieldIndex& in C++, and because python
has FieldIindex classes but no IndexP class, this allows python
iteration over output values without requiring other changes to the
C++ code.

Some properties were accessing the "components" of scalar OutputValues
by asking for outputval[0].  I think this worked only by accident.
operator[] could take an IndexP, and IndexP has a constructor that
takes a pointer to a FieldIndex, and 0 is a valid pointer to a
FieldIndex.  outputval[0] was evaluated as outputval[IndexP(nullptr)].
ScalarOutputVal::operator[] ignores its argument, so the null pointer
wasn't a problem.  Once I removed operator[](IndexP&) and replaced
it with operator[](FieldIndex&), this no longer worked.  I replaced
outputval[0] with outputval[ScalarFieldIndex()], which is clunky
but correct.

This doesn't work yet, because Equation.components() is used in
python but not yet defined.