Fixed two FieldIndex bugs.
SymmetricTensorFlux::divergenceComponents was returning the
wrong ones.  It's a question of planarity, which is assumed
to be in-plane.  I added a TODO.  Setting it in-plane makes
it consistent with the old version.

AverageField.columnNames() was using the old style field index
iterator.

Improved the debugging output in SmallMatrix::operator()(int, int).