OutputValue classes are now iterated over with conventional iterators.
Added an output operator to the PythonExportable base class, for
debugging.  Without it, it's impossible to print anything from
PythonExportableBase unless all python exportable subclasses have
an output operator.  The actual definition is in pyutils.C.

The planarity argument for Field and Flux iterators is now optional,
and defaults to ALL_INDICES.  This allows generic code to work on
Fields and Fluxes as well as OutputVals, which don't have a planarity.