Partially fixed the skeleton copying bug.
Groups are now created properly in the new Skeleton, but they're
empty.

The GenericGroupSet.nameCopy() was copying names, but not
creating groups, and was useless.  Now SkeletonContext.groupCopy()
calls GenericGroupSet.addGroups() instead, which is a new method.
GenericGroupSet.addGroup() (no s) used to add multiple groups, but
didn't work if given an OrderedSet, so I changed it to take only
a single group name, and added addGroups.