Simplified OrderedDict.
As of Python3.7, dicts preserve the order of their entries, so that
they're returned in the order in which they were added, making our
OrderedDict class nearly redundant.  The only thing it has to do
now is to add a reorder() method to dict.  OrderedSet.replace()
relied on OrderedDict.replace(), which no longer exists, but since
it didn't seem to be used anywhere else, I just rewrote OrderedSet.replace().