Minor TODOs in material.swg.
Moved the PYTHONEXPORTABLE(Property) to property.swg, where it
belongs, and %imported it into material.swg.

Deleted the TODO about having a general typemap for Enum classes.
It would only be used in two places and wouldn't add much value.
A general automatic conversion between C++ enums and Python Enums
would maybe be useful, but would be difficult to implement.