Fixes to swig files -- workarounds for a swig bug and putting some macros in the right places.
The bug is what I had been trying to work around with the NOSECONDIMPORTS
preprocessor commands, but that was a bad idea.  I now have pretty
good evidence that it's a bug and have submitted a bug report
(https://github.com/swig/swig/issues/2500). The problem is that
swig is generating code that uses pointers to classes defined in
headers that aren't necessarily included in the file.  The extra
code is harmleess because it's not called, but it prevents the
swig-generated C++ file from compiling.  The new workaround is to
add extra #include statements in the swig files.  They are all
marked "SWIGBUG" so that they can be removed easily when the bug
is fixed.

The PYTHONEXPORTABLE macro should be called in the swig file that
defines the PythonExportable subclass.  The MAKE_LISTVEC_TYPEMAPS
macro needs to be called in the file that *uses* the typemaps, not
the one that defines the class that the typemaps apply to.  This is
because the macro generates C++ code that is ignored by %import.
PYTHONEXPORTABLE only defines typemaps, which %import detects.