Fixed pdf testing problems caused by a new version of cairo, I think.
The new version was writing pdf files containing a bit of binary
data (a timestamp, maybe?) that changed from run to run, even on
the same computer.  This made all pdf file comparisons in the test
suite fail.  It was possible to make oofcanvas stick to an old
version of pdf, which made the files reproducible again, but required
them all to be rerecorded.

While figuring this out, I changed pdf_compare() in file_utils.py
so that it can compare to multiple versions of a reference file.
That turned out not to be the solution here, but might be useful
later.