Added a compilation option for whether or not to use scikit-image.
Also made the python version and swig version into cmake cache
variables, so that they can be set by ccmake instead of by emacs.

Setting the skimage flag is mostly a development convenience.  We
probably don't want it to be settable by the end user.  There were
non-scikit related improvements on the scikit-image branch which I
want to preserve.