Progress on loading numpy images from oof data files.
The image data is stored as a BytesParameter, which writes
bytes directly in binary files and converts to hex in ascii
files.  This is more compact than the old ascii format, which
wrote integers, and only slightly more unreadable.

All image data is stored using 16 bit ints for R, G, and B.  This
is compatible with the old way of storing ImageMagick images.  The
data is converted to 64 bit floats internally. We may want to change
this.

The new code should be compatible with the old format, but isn't
yet.