New aspect ratio calculation for skeleton element refinement, and new selection by aspect ratio.
The old refinement by aspect ratio didn't really do what it claimed
to do.  The new method is slower but more reliable.  It computes
the eigenvalues of the moment of inertia tensor of the nodes.  When
refining by aspect ratio, quads are subdivided only if they meet
the aspect ratio test and if their longest two edges are opposite
one another.