Added refinement rules for all bisection and trisections of triangles and quads.
This is the reworking of SnapRefine to make it use the usual
element energy to choose the best refinement.  The old
method tried to use the edge categories instead.
This new machinery should work for both Refine and SnapRefine.

I've made a first pass at adding all the refinment rules, but
I haven't actually tried to run them yet.