Fixed some more bugs in the new SnapRefine, and removed some debugging and obsolete code.
Also added some TODOs.

The check for short segments in CMicrostructure::getSegmentSections
wasn't being done if the sections were all in a line, but that was
wrong.  Now it's always done, after stair step removal.  Stairstep
removal is in an if-block so that it can be skipped if the sections
are in a line.

The code for reducing the number of marks on a segment when there
are more than the refinement rules can handle is now done in a separate
step in SnapRefine2.refinement().  It was being done too soon before.
The refinement rule set now reports how many marks it can use.