Changes related to the new segment homogeneity calculation.
SkeletonSegment.homogeneity computes the homogeneity in both
directions, and returns the smaller one.  This is what you want
to use when selecting or refining edges by homogeneity.

SkeletonSegment.homogeneity2 cmoputes the homoogeneity in both
directions, and returns both.  This is what the SkeletonInfo
toolbox reports.

Added ways of refining and selecting segments using the old homogeneity
calculation.  These only appear if OOF2 is started in debug mode.