Changed the minimum value for the min_distance parameter.
It can now be set as low as 0, but the default value is 2, and the
tooltip suggests not setting it less than that.

Also fixed problems in skeleton_periodic_test, which was trying
to use the old refinement target classes with the new snap refinement
method.  The test doesn't work yet because its reference files need
to be updated.