Made the GUI prettier by restricting the number of decimal places displayed.
The length can be set with --digits=n at startup.  It should also
get a menu item.
This wasn't crucial, but I needed something easy to do that I could
actually accomplish.  And it makes use of new f-string features.
There may be more spots that need to be changed.