The floating point precision used in the gui can now be set interactively.
This required turning runtimeflags.digits into a function.

The menu item is in the Settings menu.  It turned out that the order
of items in the Settings menu wasn't fixed -- none of the items had
an "ordering" member.  Now they all do.

Added a NonNegativeIntParameter to be used when setting the precision,
and a NonNegativeFloatParameter to be used when setting the minimum
spacing in SnapRefine.