Fixed some problems related to ReservableList.
Skeleton uses ReservableList for its nodes and elements to avoid
reallocating the lists as the Skeleton is built.  Skeleton.element_iterator
returns a ReservableListIterator.  That iterator now supports
the SkeletonIterator protocol, so that it can be queried  by a ProgressBar.

However, Skeleton.cleanUp was converting the ReservableList to a
regular list, and the standard list iterator can't be used with a
ProgressBar.  cleanUp now creates new ReservableLists, and
ReservableList.__init__ can take an optional keyword argument,
"vals", which can be either a list, tuple, or iterable object.
That should make creating the new lists reasonably efficient.