Lots of changes to refinemethod.py.
The energy function now weights the homogoneneity by element area,
so that the homogeneity index of the affected elements is maximized.
It doesn't weight the shape energy by area, so that it doesn't create
lots of small badly shaped elements.  Perhaps this change should be
made elsewhere, too.

The refinment rules in the Large set now call theVeryBest()
instead of theBetter() to choose the best provisional refinement.
The difference is that theVeryBest() generates new refinements by
subdividing the quads into two triangles, and adds them to the
set of candidates.  This means that not so many candidates have
to be coded by hand.

The Quick rules generally just create a single refinement or two,
and don't call theVeryBest().  It's not obvious when the extra
work involved in theVeryBest() is worthwhile.

Some of the rules had mistakes, which have been fixed.  There may
be more...

Not all of the possible cases are considered by the trisection
rules.  There are just too many.  I'd like to look into automating
the rule generation.

Added a TODO to refine.py about adding a way to detect when
a corner is on a pixel boundary, and adding rules to handle
those cases.  It will be a lot easier if rules are generated
automatically.

Skeleton.element_iterator now returns a SkeletonElementIterator,
which allows the progress bar to be shown in long refinement
operations.