More gui test fixes.
Since the precision of floating point numbers displayed in the GUI
is smaller now, tests that compare numbers in the gui with expected
numbers are failing, unless the tolerance is increased.  I changed
a lot of tolerances to 1.e-4 from 1.e-6.  I don't think this is
dangerous -- coding errors are likely to change results by much
larger factors.

A bigger issue is that fp_string_compare and friends in generics.py
(mainly gtkTextCompare and gtkMultiTexstCompare) were doing a pure
string comparison, instead of comparing floating point values, if
the given tolerance was 0 or None.  There are cases where we do
want a floating point comparison with tolerance=0, eg, when we want
0 and 0.0 to be the same.  So now the floating point comparison is
used when tolerance is 0, but string comparison is done when tolerance
is None.  Calls to gtkTextCompare and gtkMultiTextCompare should
be checked if they suddenly start to fail.  I didn't go back and
make changes in test files that aren't failing.