Fixed an AMR bug, and use iterators instead of lists in ElementRefinementTarget
AdaptiveMeshRefine uses the new ElementRefinementTarget class, with
a helper AMRPredicate class that invokes the AMR machinery to decide
which elements to iterate over.

amr_test.py now uses the new Refine method.  Surprisingly, the
old conservative trisection and the new quick trisection give
the same results, so the tests don't have to be rebuilt.