Minor improvements to some iterators, and a fix to a bug in MultiListWidget.
The Python implementation of ElementNodeIterator and friends is
much simpler than it was.

While testing the iterators, I made the code in ChooserListWidget
and MultiListWidget more iterator-friendly, and noticed and fixed
a bug in MultiListWidget, which had switched the order of the objects
stored in the liststore, in some circumstances.  If the bug had any
consequences, they weren't apparent.