TODOs for mesh and mesh element iterators.
ElementNodeIterator and its subclasses are used to loop over and
identify nodes in an FEMesh Element.  Converting the current C++
iterators into STL-style iterators is not entirely straightforward
because the iterators can wrap around.  I changed the TODO PYTHON3
in elementnodeiterator.h to a TODO PYTHON3 LATER and added a lot
of explanation about what has to be done.

Added some TODOs about more easily fixed iterator issues, and
fixed a typo in a comment in edge.C.