Improved the iterators in meshiterator.*.
Iteration over nodes and funcnodes in subproblems now works,
as well as iteration over meshes.

The old iterators (which weren't STL-like in C++) are still there,
but I've added OLD to their names.   I removed the NEW that was
added to the names of the new iterators in an earlier commit.

Most of the code now uses the new iterators.  There's one place
where the old iterators are still used, but I'll fix that soon.

The element iterators in meshiterator.* haven't been converted
yet.