Implemented the new iterator scheme for Elements in FEMeshes and CSubProblems.
The same machinery used for Nodes can be used for Elements, so the
names of some of the classes should be changed.

The old mesh element iterator is still there, and is used by the
interface element ("edgement") iterator.  That will be replaced
next.  The names of the old iterator classes and functions have had
"OLD" added to them temporarily.

For some mysterious reason, Element::get_index() returned a reference
to a const int, instead of a simple int.  It needs to be replaced
by Element::index() anyway, which is needed so that both Elements
and Nodes can use the same iteratory machinery.

I will change the edgement iterators next, then remove the old
iterator classes.  After that I may experiment with a different
scheme that would have less overhead but might reuse less code,
trying to access the bare stl iterators whenever possible.