Added a simpler way of iterating over FuncNodes in a FEMesh, to see if it's worth it.
A test that just looped over FuncNodes without doing anything in
the body of the loop was about 4 times faster using the simpler
scheme, which is really very simple in C++.  It's less simple in
Python, and didn't have any speed advantage over the previous
VContainer<OBJ> scheme.

Added comments about the simpler method and its pros and cons in
meshiterator.h.  It's probably worth doing.

Also changed some ints to unsigned ints.  We should be consistent
about using unsigned ints for indices.  I didn't complete the
process because I may be discarding the code that I was fixing.