Various incomplete changes to the mesh iterators.
I was trying to use bare iterators as much as possible -- ie,
not wrapping them in classes that use virtual function calls,
to make all of the iterators present the same API *and* act like
real STL and python iterators.  It may not be possible to do that.
See comments in csubproblem.h.

I am going to abandon this branch but might copy some of it onto
a new branch.