Clean up before merging with the python3-conversion branch.
This branch was started one commit before the end of the
new-mesh-iterators branch.  The final commit on that branch contains
partial work toward making some of the iterators simpler and faster.
Comments in the commit and the code explain why it was (temporarily?)
abandoned.

Renamed FEMesh::funcnode_iterator_simple to funcnode_iterator_fast.

Deleted some unused code.   Added some TODOs.