Did a TODO about the mess of advertise* functions used in problem.py.
The functions' only purpose was to install new Fields, Fluxes, and
Equations into the main namespace and to call the correct switchboard
signals.  All of that is now done more simply by python code
appended to the class's constructors by swig's %pythonappend
command.

For example, field.spy defines a function called finishConstruction
which does everything that the advertise* functions used to do. It's
added to the __init__ by  %pythonappend calls in field.swg.  It has
to be added explicitly for each subclass, unfortunately.  On the other
hand, problem.py no longer has to call advertise() for every field
instance.