Disposed of a few more TODOs.
Changed the name of CheckHeterogeneousEdges to CheckHeterogeneousSegments
for consistency with the other class names.

Deleted TODOs about whether the refinement target objects or the
iterators should be checking the refinement criterion.  It's better
done the way it is.

Greatly enlarged a TODO about the odd design of the refinement
target classes, and changed it to a TODO LATER.  There should
be a bunch of criteria with check boxes, and all of the checked
ones would be applied:  Selected, Heterogeneous, Min. Area,
etc.