Fixed a problem that was making skeleton refinement tests fail.
The skeleton refinement process generates a bunch of candidate
replacements for an element, and chooses the candidate with the
lowest energy.  Sometimes there are degenerate choices that
theoretically have exactly the same energy.  In that case, numerical
error in the energy calculation determines which one has the lowest
energy.  That numerical error can depend on the processor, so the
results may differ on different computers, which can break the
regression tests.  Tests that were passing on Linux and an Intel
iMac were failing on an M1 MacBook.

I changed the way in which refinement distinguishes between energies.
It now only chooses a new low energy candidate if the energy
difference is greater than some small number.  This means that the
first candidate seen will be chosen over later equivalent refinements.

This change broke some test files, which were regenerated.