Added tests for the non-trivial SnapNodes arguments.
Because they act on selected nodes and elements, and because
skeleton data files don't store selections, I changed the
doModify method in skeleton_basic_test.py so that you the
test data can include some extra code to be exec'd after
reading the data file and before running the modification.
This allows the tests to select the nodes and elements.