Various minor fixes.
Skeleton.element_iterator() et al can take an optional conditional
argument.

Simplified a loop in refinementtarget.py.

Adjusted a comment and removed a completed TODO in skeletonselectable.py.

Fixed the argument for setPinned (it must be a bool, not int)
and used the conditional arg for element_iterator() in
snapnode.py.