Added a new SnapNode method, which will replace the old one soon.
The new one is called SnapNode2 temporarily.

While fixing iterators, I was trying to get rid of the
TransitionPointIterator class, which was a mess.  Transition points
on skeleton edges are found better by using the new getSegmentSections
method in CMicrostructure.  There was also an old TODO about rewriting
SnapNode as a segment-based rather than element-based algorithm.
So I did that, using getSegmentSections instead of TransitionPointIterator,
and the code is much cleaner and gets better results.

I will try implementing the other long-standing TODO in SnapNodes,
which was to snap neighboring nodes in order to zip along a material
boundary, if possible.