Minor changes.
Deleted a TODO in skeleton.py and added a comment about why it was
a bad idea.

Repositioned a stray comment in refinemethod.py.

Commented out a base class method in SkelModCriterion that needs to
be provided in subclasses.  The no-op base class method could hide
a missing subclass definition.