Minor changes related to building and packaging.
Deleted the initial attempts at using cmake to create the tar
file.  It's not doing what I think it should and the old make_dist
file will work just as well.

Deleted localexclusions, which was used for packaging when we thought
that oof2 and oof3d would someday again share the same code base.

Marked oof2extutils.py as obsolete, but didn't delete it, because
it hasn't been updated yet.