Don't call SWIG_ConvertPtr within the Element constructor.
Using SWIG_ConvertPtr to extract a C++ pointer from a PyObject*
outside of a typemap is undocumented, and is failing with Python
3.11.  It was working earlier.  This workaround uses a multiargument
typemap to convert Python SkeletonElement argument to both a C++
PyObject* and a CSkeletonElement*, and passes them both to the
Element constructor.  The old way just passed in the PyObject* and
tried to extract the C++ pointer from it in the C++ Element
constructor.