Changes to get oof2 to work with python3.
(It's possible that not all of these changes are necessary.  I will
try reverting the Material+swig related stuff and see if the python
lock changes are sufficient.)

When building a python3.11 version, passing a Material in to
C++ was failing in the SWIG_ConvertPtr routine.
Assuming that it was a swig bug, I added workarounds that make
it unnecessary to do that. Material names are passed into C++, and
it's now possible to get a Material* from a material name in C++.