Trying to get cmake to generate swigruntime.h.
Currently we provide swigruntime-410.h and have no way to change
versions automatically if a different version of swig is used.  Since
the file is generated by running "swig -external-runtime", cmake
ought to be able to do it.  But I haven't found a way to get it to
work that doesn't also force all of oof2 to be rebuilt every time
"make" is run.