The swig runtime header is generated at compile time if necessary, and python and swig versions are set more easily.
When swig methods are called outside of swig-generated code, it's
necessary to include the header generated by "swig -external-runtime".
That header differs depending on the version of swig used to
generate it.   Previously we were using the header from swig 4.1
even if we were using swig 4.0.  This seemed to work, but probably
wasn't a good idea.  Now CMakeLists.txt defines a command to
generate swigruntime.h if neccesary.

Also, python and swig versions set in ccmake are chosen from
a list of options, instead of being typed by the user.