Changes to oof2.in so that setting PYTHONPATH is no longer necessary.
CMake writes the path to the installation directory into a sys.path.append line
in oof2.in.  This works properly if CMAKE_INSTALL_PREFIX is the final installation
prefix.  For staged installation, set DESTDIR when installing, eg;
   make DESTDIR=~/stow/oof2-staged
and then stow the result into /:
   cd ~/stow
   stow -t / oof2-staged
The directories inside the staging directory are relative to /, not
to ~/stow/..