Added some missing documentation.  Also a TODO in oofmenu.py and xmlmenudump.py.
The TODO is about the 'disabled' option in OOFMenuItem.  Options
are values that can be set in menu items and apply to the item and
all child items in the menu tree.  The trouble with 'disabled' is
that all of the other options are static quantities, but 'disabled'
isn't.  It makes sense the others to have their values printed in
the documentation, but the docs shouldn't report that a menu item
is disabled just because it happened to be unusable when the docs
were generated.  I added a hack to xmlmenudump.py to fix this, and
a TODO about fixing it correctly.

Also fixed a couple of python2->python3 conversion bugs in xmlmenudump.py.