SymTensor indices can now be looped over in python.
A getSymTensorIterator function had been declared but not used or
defined.  I replaced it with getSymTensorComponents().  It's used
in StrainType.components(), which is used when the GUI lists the
strain components for defining outputs, for example.  Somehow this
hadn't been tested for.

I bumped the version number to 2.3.2 and started a changelog
section for it.