Fixed an Output bug, and added tests for it.
RangeOutput wasn't working with ElementSegmentSampleSet.
The problem was that that SampleSet returns more than one value,
which Range wasn't expecting.  Now expandRange is a Sample method,
so that different kinds of Samples can do it differently.

Deleted initRange, since it wasn't necessary when expandRange
can handle initial None values.

Added operator< to ScalarOutputVal so that the range computation
can be done easily.