Fixed some bugs in cloning graphics windows, and rewrote the GfxSettings class.
Graphics settings weren't being passed to cloned windows properly,
and sometimes the cloning process would crash.  I'm not 100% sure
that I've fixed the crashes, but the parameter initialization
and copying is working much better.

Instead of storing settings as class-level data in the GfxSettings
objects, and distinguishing between instance data and class data,
now there's a separate default_settings dict that stores the
default values.  In order to create new or cloned graphics windows
with the correct settings, a GfxSettings object (or None) is passed
in to the GhostGfxWindow or GfxWindowBase constructor.  If the
the passed-in settings is None, a new GfxSettings object is
constructed from default_settings.