Menus and RegisteredClasses now can have option "xrefs" for cross linking the manual.
Manual pages for OOFMenuItems and RegisteredClasses will generate
a "See Also" section with links to other parts of the manual if
the menu items, registrations, or registered classes have an
"xrefs" data member.  xrefs is a list containing the xml ids
of some other parts of the manual or (id, text) pairs if the
default link text isn't sufficient.

xrefs in a menu will be displayed for all menu items in the menu,
recursively.  xrefs in a registered class will be displayed in the
base class and all subclasses.