Fixed some errors in Rationalize.
The methods in rationalsharp.py are more careful about ensuring
that new midpoints of horizontal segments have the exact y values
as the end points, and the same for vertical segments and x values.
This ensures that new boundary nodes actually are on the boundaries,
and therefore have the correct mobilities assigned.

Skeleton.sanity_check is much faster.  It now copies nodes and
elements into dicts for faster lookup.  It's no longer really
necessary to have a quick sanity check that skips all the slow
parts.  It also checks that the nodes on the exterior edges have
the correct x or y components.

The call to skel.cleanUp was indented too far and was being done
after each Element was examined, insetad of at the end of the loop.
It's been at the wrong place since 2.2.0.

Added a separate __repr__ for ProvsionalElement because I wanted
to change the base class (SkeletonElementBase) repr to something
that ProvisonalElement couldn't handle, and it was easier to do
that than to change the other derived classes.  It's just for
debugging anyway.