Partially fixed problems with animating graphics windows.
findProgress() was raising an exception because it only looks for
Progress objects that were constructed on the current thread.
Because animation is run on a timeout callback on the main thread,
it's necessary to use findProgressByID instead.

Animation is still not as smooth as it should be for high frame
rates.