Fixed a bug in command line argument handling.
process_inline_options() used to import an oof2 module so that it
could set a flag.  But that module imported another one that built
an OOF menu, so that menu was built before the options were fully
processed.  That was messing up the xml generation for the manual.
Now there are lots of comments about doing things in the right
order.