Better organization in the xml files for menus.
Rewrote dumpMenu().  The old way intermixed submenus and commands
in the listing and also created pages for command listings that
weren't accessible at all.  There wasn't much sense in what the
Next and Prev buttons pointed to.

Also changed variables named 'file' to 'phile', because although
'file' isn't a reserved word in Python, emacs thinks it is,
and was coloring it as such.