Added an 'alphabetize' option to OOFMenuItem, and an emptyDiscussion object.  Also more docs.
If an OOFMenuItem is created with the optional keyword arg 'alphabetize=False'
then its items won't be alphabetized in the manual.  The default value
is to alphabetize in the manual.

The tutorial menu is not alphabetized, so that the listing in the
manual follows the natural order.

A menu item or registered class can set its discussion
to emptyDiscussion if there's really nothing to say.
If the discussion is simply omitted, a MISSING message
is printed in the manual.