More documentation.  Fixed some doc generation bugs.
A bug in two places in propertyregistration.py was causing documentation
for OOF.Property and OOF.LoadData.Property to not show up in the
manual.

Also deleted some old oof3d code because it was getting in the way.

Changed the generic info printed for a couple of parameter types
in parameter.py because it was confusing.  The
ListOfTuplesOfFloatsParameter and ListOfTuplesOfIntsParameter classes
don't require the tuples to be all the same length, but they don't
need to say that in the manual.  Some arguments had been described like
"A list of tuples of two ints. The tuples do not have to be all the
same length".