Added a --retries option to guitests.
Some of the tests fail intermittently, at least on macOS. This
seems to be a problem in the test machinery, not in OOF2 itself.
The tests eventually pass if repeated enough times.  I got tired
of restarting them manually, so now you can use
   oof2-guitest --retries=N
or
  python guitests.py --retries=N
to automatically repeat failed tests up to N times.