Better documentation for the various undo and redo commands.
The old versions talked about pushing and popping selections on a
stack.  The problem is that it's not a stack, and going into detail
wasn't helping to explain it.  I just simplified the discussions.

Also added a TODO about not needed separate menu items for undo and
redo from the main window and from the graphics toolboxes.