Made the skeleton selection tools consistent in the way they handle lists of points.
Some of the tools, like OOF.Graphics_n.Toolbox.Select_Element.Ellipse
and its friends, have an argument that is a list of points. Most of
them use two points from the list.  Some of them used the first two,
and some used the first and last.  Now they all use the first and
last.

This change probably affects nobody, but it makes the documentation
easier to write.  The GUI generates lists with only two points when
only two are needed, so the second and last points were identical.