Tweaked the "See Also" listings in the generated manual pages.
It now uses an itemizedlist, which is much more readable when there
are more than a handful of cross references.  It also sorts the
list, which doesn't quite work, because it sorts on the docbook
ids, which in most cases is not the same thing as sorting on the
text in the list items.  There's a long comment about maybe doing
it better.  Doing it actually correctly would require modifying
docbook.