Better sorting of xrefs in the See Also sections of the manual.
Xrefs to targets in the hand-written chapters are sorted in
the order that their targets appear in the manual.  Mostly this
means in numerical order of their Section number.  Targets in
the reference pages generated by xmlmenudump are sorted alphabetically
within their category (menu item, regsitered class, etc).

This machinery relies on changes to xmldump.py in the MAN_OOF2
directory.  It also makes it inconvenient to use an OOF2 menu item
to generate the manual.  Since there's no real need for that, I
deleted the menu command for it, and for dumping a text (non-xml)
version of the reference pages.  I don't think that was very useful
anyway.