Reorganized the directory of extension templates.
Got rid of one directory level.  It will be recreated when a template
is instantiated by oof2-extension-setup, but there's no need to
clutter up the source with it.

Renamed all source files to oof2template.*.  This will simplify the
code in oof2-extension-setup.  The different templates are still
distinguished by their directory name.

Didn't change teh ramberg-osgood subdirectory.  It doesn't really
belong in this directory, and we'll have to find another home for
it, perhaps with the internal extensions (OOF2/SRC/EXTENSIONS).