More reorganization.  Added an "archived" directory for old code.
Some old projects have been abandoned but are still in the code
base.  I don't want to delete them entirely, because someone
might want to continue them.  Putting them all in a "archived"
subdirectory makes them easier to ignore until they're revived.

This was motivated the the mods I'm making to the way
extensions are built.  All of the header files are going to
be installed in <prefix>/include/oof2, so that extensions
can be built even if the oof2 source code isn't available,
but I don't want to install too many obsolete or unused headers.